Added `v2/decks/{deck_id}` to available endpoints to read latest public snapshot
Also allows viewing saved copy of the source deck (for the owner) and returns snapshots (based on access permissions).

Additionally refactored the decks services into the existing deck service.
